Refactor python scripts
